# docs/api/semaphore.openapi.yaml
openapi: 3.1.0
info:
  title: KApp Semaphore API
  version: 1.0.0
  summary: Academic catalog and student pensum progress (semáforo) for Fundación Universitaria Konrad Lorenz.
  description: |
    The Semaphore service owns both the **academic catalog** (programs and pensums) and each
    student's **pensum progress**, because at Konrad Lorenz the pensum *is* the semáforo.

    A *semáforo* renders a whole degree plan — the *malla curricular*, also called the *pensum* —
    as a grid: columns are levels (semesters, `1..9`), rows are knowledge areas, and every pensum
    item is coloured by its status. Hence "traffic light".

    ## Conventions used across this API

    - **Grades** are integers on the university's `0..50` scale (mirrors the `final_grade`
      constraint in the KApp database). There are no decimals. The passing threshold is applied by
      the registrar, not by this service; this service only stores and validates the range.
    - **Periods** are the university's own academic-period strings: four digits for the year
      followed by `1` or `2` for the semester (`^\d{4}[12]$`). `"20262"` is the second semester of
      2026.
    - **Hours.** A course carries `weeklyHours` (contact hours per week) and, derived from it,
      `totalHours` for the full 16-week semester. The invariant is `weeklyHours * 16 == totalHours`.
      At *pensum* and *area* level, `totalHours` and `hours` are sums of **weekly** hours across
...
      `pensumItemCode`.
...

    ## Authorization

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /api/semaphore/me/plans:
    get:
      # Authorization as the code enforces it, not as prose describes it.
      x-roles: [ROLE_STUDENT]
      operationId: listMyAcademicPlans
      summary: List the student's academic plans
      description: |
        Requires `ROLE_STUDENT`. Returns the caller's own plans, the primary one first.

        A plan is a layer over the pensum, never a copy of it. The pensum stays the
        immutable *Semáforo Original*; a plan records only the courses the student has moved,
        which is why `placements` is usually short and often empty.
      tags: [Academic Plans]
      responses:
        '200':
          description: The caller's plans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AcademicPlan'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      # Authorization as the code enforces it, not as prose describes it.
      x-roles: [ROLE_STUDENT]
      operationId: createMyAcademicPlan
      summary: Create an academic plan
      description: |
        Requires `ROLE_STUDENT`. The plan starts empty: every course sits where the pensum puts
        it until the student moves one.

        The first plan a student creates for a pensum becomes their primary automatically.
        A student may hold at most 10 plans per pensum — enough for "what if I take Cálculo in
        the summer", far short of an unbounded collection growing under one account.
      tags: [Academic Plans]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AcademicPlanRequest'
      responses:
        '201':
          description: The created plan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcademicPlan'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/semaphore/me/plans/{planId}:
    parameters:
      - $ref: '#/components/parameters/PlanIdPath'
    get:
      # Authorization as the code enforces it, not as prose describes it.
      x-roles: [ROLE_STUDENT]
      operationId: getMyAcademicPlan
      summary: Get one academic plan
      description: |
        Requires `ROLE_STUDENT`. `404` for a plan belonging to somebody else, deliberately: a
        `403` would confirm that the identifier exists.
      tags: [Academic Plans]
      responses:
        '200':
          description: The plan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcademicPlan'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      # Authorization as the code enforces it, not as prose describes it.
      x-roles: [ROLE_STUDENT]
      operationId: updateMyAcademicPlan
      summary: Rename a plan or make it the primary one
      description: |
        Requires `ROLE_STUDENT`. Both fields are optional; sending neither is a `400`.

        Setting `primary` to `true` demotes whichever plan currently holds it for the same
        pensum — there is exactly one primary per student and pensum. Setting it to `false` is
        rejected: a student would be left with none, and something has to be the plan the app
        opens on.
      tags: [Academic Plans]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AcademicPlanUpdate'
      responses:
        '200':
          description: The updated plan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcademicPlan'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      # Authorization as the code enforces it, not as prose describes it.
      x-roles: [ROLE_STUDENT]
      operationId: deleteMyAcademicPlan
      summary: Delete an academic plan
      description: |
        Requires `ROLE_STUDENT`. Deleting the primary plan promotes the next one by creation
        date; deleting the last plan for a pensum is allowed and simply leaves the student with
        the pensum as published.
      tags: [Academic Plans]
      responses:
        '204':
          description: Deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/semaphore/me/plans/{planId}/placements/{code}:
    parameters:
      - $ref: '#/components/parameters/PlanIdPath'
      - $ref: '#/components/parameters/CourseCodePath'
    put:
      # Authorization as the code enforces it, not as prose describes it.
      x-roles: [ROLE_STUDENT]
      operationId: placeCourseInMyPlan
      summary: Move a course to a different level
      description: |
        Requires `ROLE_STUDENT`. Idempotent: moving a course already at that level returns the
        plan unchanged.

        **Moving a course does not make it eligible.** `GET /api/semaphore/me/eligible` is
        computed from prerequisites actually approved, and it ignores plans entirely. Planning
        is not approving, and a plan that could unlock a course by dragging it would be a
        semáforo that lies.

        There is **no cap on how many courses may sit at one level**. Students take more or
        fewer than the nominal six all the time, and a limit here would refuse a real timetable.

        `400` when the course is not in the pensum, or when `plannedLevel` is outside 1–12.
      tags: [Academic Plans]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlacementRequest'
      responses:
        '200':
          description: The plan, including the new placement
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcademicPlan'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      # Authorization as the code enforces it, not as prose describes it.
      x-roles: [ROLE_STUDENT]
      operationId: resetCourseInMyPlan
      summary: Return a course to the level the pensum gives it
      description: |
        Requires `ROLE_STUDENT`. Removes the placement, so the course goes back to its published
        level. `404` when the plan has no placement for that course — it was never moved.
      tags: [Academic Plans]
      responses:
        '200':
          description: The plan, without that placement
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcademicPlan'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiError:
      type: object
      required: [timestamp, status, error, message, path]
      properties:
        timestamp:
          description: >-
            When the service produced this error, in UTC. Not when the request was sent.
          type: string
          format: date-time
        status:
          description: >-
            The HTTP status, repeated in the body so a logged response is self-contained.
          type: integer
          example: 400
        error:
          description: >-
            The status's standard name - `Bad Request`, `Conflict`. For developers, not for users.
          type: string
          example: Bad Request
        message:
          description: >-
            One sentence a person can act on. This is the line worth surfacing in a UI.
          type: string
        path:
          description: >-
            The path that produced the error, so a log line identifies itself.
          type: string
        details:
          description: >-
            Field-by-field reasons, when there are any. `field` names what was wrong and `issue` says what
            about it - this is what a form binds its per-field errors to.
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              issue: { type: string }
...
    AcademicPlan:
      type: object
      description: |
        A student's own arrangement of a pensum.

        `placements` carries **only the courses that moved**. Anything absent is taken where the
        pensum puts it. That keeps the document small, makes the outline the client draws in
        the original column fall out for free — it knows both levels — and means a change to the
        pensum reaches every plan without rewriting any of them.
      required: [id, name, pensumCode, primary, placements, createdAt, updatedAt]
      properties:
        id:
          type: string
          maxLength: 64
          example: 66f0a1c2e4b0a91d3c5e7f21
        name:
          type: string
          maxLength: 60
          example: Mi plan
        pensumCode:
          description: >-
            The pensum this plan layers over. The plan stores only the moves; everything else is read from
            the pensum.
          type: string
          maxLength: 20
          example: "1015"
        primary:
          type: boolean
          description: The plan the app opens on. Exactly one per student and pensum.
          example: true
        placements:
          type: array
          description: Only the courses the student moved.
          items:
            $ref: '#/components/schemas/Placement'
        createdAt:
          description: >-
            When the student created the plan.
          type: string
          format: date-time
        updatedAt:
          description: >-
            When the plan last changed - a move, a rename, or being made primary.
          type: string
          format: date-time
    Placement:
      type: object
      description: One course pinned to a level other than the pensum's.
      required: [code, plannedLevel]
      properties:
        code:
          type: string
          maxLength: 30
          description: |
            The course `code`, or an elective slot's `pensumItemCode` — the same identifier
            `PUT /api/semaphore/me/courses/{code}` accepts.
          example: "17085"
        plannedLevel:
          type: integer
          minimum: 1
          maximum: 12
          description: The level the student intends to take it in.
          example: 7
    AcademicPlanRequest:
      type: object
      required: [name, pensumCode]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 60
          example: Mi plan
        pensumCode:
          type: string
          maxLength: 20
          description: Must be an existing pensum; `404` otherwise.
          example: "1015"
      additionalProperties: false
    AcademicPlanUpdate:
      type: object
      description: |
        At least one field must be present. An empty object is a `400` rather than a silent
        no-op, so a client bug does not look like a successful save.
      minProperties: 1
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 60
          example: Plan de verano
        primary:
          type: boolean
          description: |
            Only `true` is accepted. Demoting a plan without promoting another would leave the
            student with no primary at all.
          enum: [true]
      additionalProperties: false
    PlacementRequest:
      type: object
      required: [plannedLevel]
      properties:
        plannedLevel:
          description: >-
            The level the student wants to take this course in. Planning only: it does not change what the
            prerequisite rules allow.
          type: integer
          minimum: 1
          maximum: 12
          example: 7
      additionalProperties: false
...
  parameters:
...
    CourseCodePath:
      name: code
      in: path
      required: true
      description: |
        Identifier of the pensum item. The course `code` for a fixed course, or — because an
        elective slot has no `code` — the slot's `pensumItemCode`.
      schema:
        type: string
        maxLength: 30
      example: "48022"
...
    PlanIdPath:
      name: planId
      in: path
      required: true
      description: Identifier of one of the caller's own academic plans.
      schema:
        type: string
        maxLength: 64
      example: 66f0a1c2e4b0a91d3c5e7f21
...
  responses:
    BadRequest:
      description: Validation failed
      content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } }
    Unauthorized:
      description: Missing or invalid token
      content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } }
    Forbidden:
      description: Authenticated but not allowed
      content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } }
    NotFound:
      description: Resource not found
      content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } }
    Conflict:
      description: Conflicts with current state
      content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } }